A-CodingStandard: Align code to SE-EDU/Google style; Checkstyle clean
