+++ config/checkstyle/checkstyle.xml
@@ -1,0 +1,88 @@
+<?xml version="1.0"?>
+<!DOCTYPE module PUBLIC
+    "-//Checkstyle//DTD Checkstyle Configuration 1.3//EN"
+    "https://checkstyle.org/dtds/configuration_1_3.dtd">
+
+<!-- Google-style baseline, adjusted for 120-char max line length
+     and common SE-EDU expectations (no star imports, K&R braces, etc.). -->
+<module name="Checker">
+  <!-- Charset -->
+  <property name="charset" value="UTF-8"/>
+
+  <!-- Line length: hard limit 120 -->
+  <module name="LineLength">
+    <property name="max" value="120"/>
+    <property name="ignorePattern" value="^package |^import |^\s*//|https?://"/>
+  </module>
+
+  <!-- Disallow tabs -->
+  <module name="FileTabCharacter">
+    <property name="eachLine" value="true"/>
+  </module>
+
+  <!-- Package name lowercase -->
+  <module name="Regexp">
+    <property name="format" value="^package [a-z0-9_.]+;\s*$"/>
+    <property name="message" value="Package names should be all lowercase."/>
+    <property name="ignoreCase" value="false"/>
+    <property name="illegalPattern" value="false"/>
+  </module>
+
+  <module name="TreeWalker">
+    <!-- Imports -->
+    <module name="AvoidStarImport"/>
+    <module name="UnusedImports"/>
+    <module name="ImportOrder">
+      <property name="ordered" value="true"/>
+      <property name="separated" value="true"/>
+      <property name="groups" value="java,javax,org,com"/>
+      <property name="option" value="top"/>
+      <property name="sortStaticImportsAlphabetically" value="true"/>
+      <property name="sortImportsInGroupAlphabetically" value="true"/>
+    </module>
+
+    <!-- Braces & structure -->
+    <module name="LeftCurly">
+      <property name="option" value="eol"/>
+      <property name="ignoreEnums" value="false"/>
+    </module>
+    <module name="RightCurly">
+      <property name="option" value="same"/>
+    </module>
+    <module name="NeedBraces"/>
+
+    <!-- Whitespace and wrapping -->
+    <module name="WhitespaceAround"/>
+    <module name="WhitespaceAfter"/>
+    <module name="NoWhitespaceBefore"/>
+    <module name="MethodParamPad"/>
+    <module name="OperatorWrap">
+      <property name="option" value="eol"/>
+    </module>
+    <module name="EmptyLineSeparator">
+      <property name="allowNoEmptyLineBetweenFields" value="true"/>
+    </module>
+
+    <!-- Modifiers order -->
+    <module name="ModifierOrder"/>
+
+    <!-- Naming -->
+    <module name="PackageName"/>
+    <module name="TypeName"/>
+    <module name="MethodName"/>
+    <module name="ParameterName"/>
+    <module name="LocalVariableName"/>
+    <module name="MemberName"/>
+    <module name="ConstantName"/>
+
+    <!-- Javadoc basic checks for public types/methods only -->
+    <module name="JavadocType"/>
+    <module name="JavadocMethod">
+      <property name="scope" value="public"/>
+      <property name="allowMissingParamTags" value="true"/>
+      <property name="allowMissingThrowsTags" value="true"/>
+      <property name="allowMissingReturnTag" value="true"/>
+      <property name="allowMissingJavadoc" value="false"/>
+    </module>
+  </module>
+</module>
