+++ src/main/java/v/command/ListCommand.java
@@ -1,0 +1,24 @@
+package v.command;
+
+import v.task.TaskList;
+import v.ui.Ui;
+import v.storage.Storage;
+
+/**
+ * Command to list all tasks.
+ */
+public class ListCommand extends Command {
+    /**
+     * Executes this command by displaying all tasks in the list.
+     *
+     * @param tasks   The task list to display.
+     * @param ui      The UI to render the list.
+     * @param storage The storage (unused).
+     * @return False to continue running the program.
+     */
+    @Override
+    public boolean execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.showTaskList(tasks);
+        return false;
+    }
+}
