+++ v_data.txt
@@ -2,4 +2,5 @@ T | 0 | borrow book
 E | 0 | project meeting | Mon 2pm | 4pm
 T | 0 | task1
 T | 0 | task2
-T | 0 | task3
+D | 1 | booktok | 2025-08-08
+T | 0 | booking
